AI Self-Healing Test Automation Framework/requirements.txt
selenium==4.15.2
webdriver-manager==4.0.1
pytest==7.4.3
pytest-html==4.1.1
openpyxl==3.1.5
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
openai==1.3.1
pydantic==2.5.0
